Compilation warnings fixes

+++ src/xpiks-qt/Common/basickeywordsmodel.h
@@ -121,7 +121,7 @@ namespace Common {
         void unlockKeywords() { m_KeywordsLock.unlock(); }
 
     public:
-        virtual QString retrieveKeyword(int wordIndex);
+        virtual QString retrieveKeyword(size_t wordIndex);
         virtual QStringList getKeywords();
         virtual void setKeywordsSpellCheckResults(const std::vector<std::shared_ptr<SpellCheck::SpellCheckQueryItem> > &items);
         virtual std::vector<KeywordItem> retrieveMisspelledKeywords();
